+++ piecrust/src/session.rs
@@ -17,7 +17,7 @@ use dusk_wasmtime::{
     Engine, LinearMemory, MemoryCreator, MemoryType, Module as WasmtimeModule,
 };
 use piecrust_uplink::{
-    ARGBUF_LEN, CONTRACT_ID_BYTES, ContractId, EnrichedEvent, SCRATCH_BUF_BYTES,
+    ARGBUF_LEN, CONTRACT_ID_BYTES, ContractId, Event, SCRATCH_BUF_BYTES,
 };
 use rkyv::ser::Serializer;
 use rkyv::ser::serializers::{
@@ -111,7 +111,7 @@ struct SessionInner {
     buffer: Vec<u8>,
 
     feeder: Option<mpsc::Sender<Vec<u8>>>,
-    events: Vec<EnrichedEvent>,
+    events: Vec<Event>,
 
     #[cfg(feature = "call-hook")]
     call_hook: Option<CallHook>,
@@ -679,7 +679,7 @@ impl Session {
         self.inner().contract_session.memory_pages(contract)
     }
 
-    pub(crate) fn push_event(&mut self, event: EnrichedEvent) {
+    pub(crate) fn push_event(&mut self, event: Event) {
         self.inner_mut().events.push(event);
     }
 
@@ -1038,7 +1038,7 @@ pub struct CallReceipt<T> {
     pub gas_limit: u64,
 
     /// The events emitted during the execution of the call.
-    pub events: Vec<EnrichedEvent>,
+    pub events: Vec<Event>,
     /// The call tree produced during the execution.
     pub call_tree: CallTree,